+++ src/standard/sprite/sprite_system.test.ts
@@ -23,7 +23,7 @@ import SystemEntity from "../../system/system_entity";
 import FakeEntity from "../../fake/entity";
 import Transform from "../transform/transform";
 import Sprite from "./sprite";
-import Polygon from "../shape/polygon";
+import Polygon from "../../shape/polygon";
 import Reactor from "../../fake/reactor";
 import Component from "../../component/component";
 import System from "../../system/system";